Move stream file
